Refactor too use a bit less signal magic
